API_strings: adding scripts to dump API strings
This PR adds a script that will dump a hex of the admin.macaroon and a
cert thumb print of the tls.cert. Both these strings are needed for
connection to the LND Node via REST. This was made specifically in mind
for connecting a external LND Node to BTCPayServer but can also be used
for connecting to Lighting Apps.